+++ src/common/nc_version.pas
@@ -11,10 +11,6 @@ const
     c_engine_version = '0.1.0';
     c_ipc_protocol_major = 1;
     c_ipc_protocol_minor = 0;
-    c_windows_baseline_commit =
-        'f3510badeabd223c6535d81bc4255bb037271e1d';
-    c_lexicon_baseline_commit =
-        '9a951a4a118505c0e9964f0f2ddf618940bed355';
 
 implementation